MPC_Allocator: Improve RFQ by using atomic queue instead of protecting it with spinlock. MPC_TestSuite: Remove manual compilation of allocator tests, use already existing cmake scripts for standlone mode.

+++ mpc/MPC_Allocator/src/sctk_alloc_rdtsc.h
@@ -48,8 +48,8 @@
 #endif
 
 /************************* FUNCTION ************************/
-static __inline__ unsigned long sctk_alloc_rdtsc() {
-    unsigned long t;
+static __inline__ unsigned long long sctk_alloc_rdtsc() {
+    unsigned long long t;
     __sctk_rdtscll(t);
     return t;
 }
